utils: Removed ParallelObject
